+++ src/constants.rs
@@ -1,3 +1,3 @@
-pub const DEFAULT_MAX_CONTENT_LENGTH: u64 = 1024 * 1024 * 50; // 50 megabytes
+pub const DEFAULT_MAX_CONTENT_LENGTH: usize = 1024 * 1024 * 50; // 50 megabytes
 pub const DEFAULT_AUTH_TOKEN_ENDPOINT: &str = "https://tokens.indieauth.com/token";
 pub const DEFAULT_AUTH_ENDPOINT: &str = "https://indieauth.com/auth";
